+++ src/main/java/com/supermartijn642/chunkloaders/ChunkLoaders.java
@@ -1,8 +1,6 @@
 package com.supermartijn642.chunkloaders;
 
 import com.supermartijn642.chunkloaders.capability.ChunkLoadingCapability;
-import com.supermartijn642.chunkloaders.capability.ClientChunkLoadingCapability;
-import com.supermartijn642.chunkloaders.capability.ServerChunkLoadingCapability;
 import com.supermartijn642.chunkloaders.generators.*;
 import com.supermartijn642.chunkloaders.packet.*;
 import com.supermartijn642.core.CommonUtils;
@@ -10,24 +8,13 @@ import com.supermartijn642.core.item.CreativeItemGroup;
 import com.supermartijn642.core.network.PacketChannel;
 import com.supermartijn642.core.registry.GeneratorRegistrationHandler;
 import com.supermartijn642.core.registry.RegistrationHandler;
-import net.minecraft.core.Direction;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.Tag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.level.Level;
-import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.common.capabilities.*;
-import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.common.world.ForgeChunkManager;
-import net.minecraftforge.event.AttachCapabilitiesEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.neoforge.common.world.chunk.RegisterTicketControllersEvent;
+import net.neoforged.neoforge.common.world.chunk.TicketController;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 
 /**
  * Created 7/7/2020 by SuperMartijn642
@@ -35,17 +22,17 @@ import javax.annotation.Nullable;
 @Mod("chunkloaders")
 public class ChunkLoaders {
 
-    public static Capability<ChunkLoadingCapability> CHUNK_LOADING_CAPABILITY = CapabilityManager.get(new CapabilityToken<>() {
-    });
-
     public static final Logger LOGGER = LogManager.getLogger("chunkloaders");
     public static final PacketChannel CHANNEL = PacketChannel.create("chunkloaders");
     public static final CreativeItemGroup GROUP = CreativeItemGroup.create("chunkloaders", ChunkLoaderType.ADVANCED::getItem);
 
-    public ChunkLoaders(){
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(this::init);
-        FMLJavaModLoadingContext.get().getModEventBus().addListener(this::registerCapabilities);
-        MinecraftForge.EVENT_BUS.addGenericListener(Level.class, this::attachCapabilities);
+    public static final TicketController TICKET_CONTROLLER = new TicketController(
+        new ResourceLocation("chunkloaders", "chunks"),
+        (level, ticketHelper) -> ChunkLoadingCapability.get(level).castServer().onLoadLevel(ticketHelper)
+    );
+
+    public ChunkLoaders(IEventBus eventBus){
+        eventBus.addListener(this::registerTicketController);
 
         CHANNEL.registerMessage(PackedChunkLoaderAdded.class, PackedChunkLoaderAdded::new, true);
         CHANNEL.registerMessage(PackedChunkLoaderRemoved.class, PackedChunkLoaderRemoved::new, true);
@@ -61,43 +48,8 @@ public class ChunkLoaders {
         registerGenerators();
     }
 
-    public void init(FMLCommonSetupEvent e){
-        // Set the chunk loading callback
-        ForgeChunkManager.setForcedChunkLoadingCallback("chunkloaders", (level, ticketHelper) -> {
-            level.getCapability(CHUNK_LOADING_CAPABILITY).ifPresent(capability -> capability.castServer().onLoadLevel(ticketHelper));
-            level.getCapability(LegacyChunkLoadingCapability.TRACKER_CAPABILITY).ifPresent(capability -> capability.onLoadLevel(ticketHelper));
-        });
-    }
-
-    public void registerCapabilities(RegisterCapabilitiesEvent e){
-        // Register the chunk loading capability
-        e.register(ChunkLoadingCapability.class);
-
-        // Register the legacy capability
-        LegacyChunkLoadingCapability.register(e);
-    }
-
-    public void attachCapabilities(AttachCapabilitiesEvent<Level> e){
-        Level level = e.getObject();
-        LazyOptional<ChunkLoadingCapability> tracker = LazyOptional.of(() -> level.isClientSide ? new ClientChunkLoadingCapability(level) : new ServerChunkLoadingCapability(level));
-        e.addCapability(new ResourceLocation("chunkloaders", "chunk_loading_capability"), new ICapabilitySerializable<>() {
-            @Nonnull
-            @Override
-            public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side){
-                return cap == CHUNK_LOADING_CAPABILITY ? tracker.cast() : LazyOptional.empty();
-            }
-
-            @Override
-            public Tag serializeNBT(){
-                return tracker.map(ChunkLoadingCapability::write).orElse(null);
-            }
-
-            @Override
-            public void deserializeNBT(Tag nbt){
-                tracker.ifPresent(cap -> cap.read((CompoundTag)nbt));
-            }
-        });
-        e.addListener(tracker::invalidate);
+    public void registerTicketController(RegisterTicketControllersEvent e){
+        e.register(TICKET_CONTROLLER);
     }
 
     private static void register(){