+++ src/main/java/com/supermartijn642/chunkloaders/capability/ChunkLoadingEventHandler.java
@@ -3,9 +3,9 @@ package com.supermartijn642.chunkloaders.capability;
 import com.supermartijn642.chunkloaders.ChunkLoaders;
 import com.supermartijn642.chunkloaders.packet.PacketFullCapabilityData;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraftforge.event.entity.player.PlayerEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 
 /**
  * Created 26/06/2022 by SuperMartijn642