+++ src/main/java/com/supermartijn642/chunkloaders/extensions/ChunkLoadersLevel.java
@@ -1,0 +1,11 @@
+package com.supermartijn642.chunkloaders.extensions;
+
+import com.supermartijn642.chunkloaders.capability.ChunkLoadingCapability;
+
+/**
+ * Created 20/02/2023 by SuperMartijn642
+ */
+public interface ChunkLoadersLevel {
+
+    ChunkLoadingCapability getChunkLoadingCapability();
+}