+++ changelog.md
@@ -1,2 +1,2 @@
-### Chunk Loaders 1.2.8a
-- Initial release of Chunk Loaders for Minecraft 1.20.3 & 1.20.4
+### Chunk Loaders 1.2.8
+- Initial release of Chunk Loaders for NeoForge